perf: lazy load recharts components
