+++ web/src/app/[locale]/(dashboard)/dashboard/agent/page.tsx
@@ -3,7 +3,17 @@
 import { useEffect, useRef, useState } from 'react';
 import { useChat } from '@ai-sdk/react';
 import { DefaultChatTransport, type UIMessage } from 'ai';
-import { AgentChart, type AgentChartSpec } from '@/components/agent/agent-chart';
+import type { AgentChartSpec } from '@/components/agent/agent-chart';
+import dynamic from 'next/dynamic';
+import { Skeleton } from '@/components/ui/skeleton';
+const AgentChart = dynamic(
+  () =>
+    import('@/components/agent/agent-chart').then((m) => m.AgentChart),
+  {
+    ssr: false,
+    loading: () => <Skeleton className="h-64 w-full" />,
+  }
+);
 import { useFeatureGate } from '@/hooks/use-feature-gate';
 import { Link } from '@/i18n/navigation';
 import { Button, buttonVariants } from '@/components/ui/button';
